+++ th128.cpp
@@ -18,9 +18,9 @@ void th128decode(char *file, unsigned char * buffer, unsigned int flength) {
 	unsigned char *rawdata = &buffer[0x24], *stagedata, *replaydata, *fpsdata, *decodedata;
 	unsigned int length = *((unsigned int*) (&buffer[0x1c]));
 	unsigned int dlength = *((unsigned int*) (&buffer[0x20]));
-	unsigned short keys[16] =
+	unsigned int keys[16] =
 			{ 0xf0a1, 0xfca1, 0xfda1, 0xfda1, 0xfba1, 0x49a8, 0x4ca8, 0x49a8, 0xfaa1, 0x4aa8, 0x4ba8, 0x4aa8, 0xfba1, 0x49a8, 0x4ca8, 0x49a8 };
-	unsigned short skey[61];
+	unsigned int skey[61];
 	sprintf(frec, "%s.txt", file);
 	sprintf(fraw, "%s.raw", file);
